move RawValue to its own submodule
